updated base checker
