+++ demisto_sdk/commands/lint/commands_builder.py
@@ -146,10 +146,12 @@ def build_pylint_command(files: List[Path], support_level: str = "") -> str:
     """
     support_levels = {
         'base': 'base_checker',
-        'certified partner': 'certified_partner_checker',
-        'community': 'community_checker',
-        'partner': 'partner_checker',
-        'xsoar': 'xsoar_checker'
+        'community': 'base_checker,community_level_checker',
+        'partner': 'base_checker,community_level_checker,partner_level_checker',
+        'certified partner': 'base_checker,community_level_checker,partner_level_checker,'
+                             'certified_partner_level_checker',
+        'xsoar': 'base_checker,community_level_checker,partner_level_checker,certified_partner_level_checker,'
+                 'xsoar_level_checker '
     }
     command = "python -m pylint"
     # Excluded files
@@ -157,8 +159,7 @@ def build_pylint_command(files: List[Path], support_level: str = "") -> str:
     # Prints only errors
     command += " -E"
     # Load plugins
-    command += f" --load-plugins {support_levels.get('base')},{support_levels.get(support_level)}" if support_levels.get(
-        support_level) else ""
+    command += f" --load-plugins {support_levels.get(support_level)}" if support_levels.get(support_level) else ""
     # Disable specific errors
     command += " -d duplicate-string-formatting-argument"
     # List of members which are set dynamically and missed by pylint inference system, and so shouldn't trigger
