+++ demisto_sdk/commands/lint/resources/pylint_plugins/partner_level_checker.py
